% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{getIntersectionMae}
\alias{getIntersectionMae}
\title{Get the intersection for MAE -of the ASE calling, SNP annotation1
 and SNP annotation2}
\usage{
getIntersectionMae(df_aseSnpAnn1, df_aseSnpAnn2)
}
\arguments{
\item{df_aseSnpAnn1}{dataframe object containing the intersection
of the ASE calling dataset and SNP annotation1 -the output of
getIntersection()}

\item{df_aseSnpAnn2}{dataframe object containing the intersection
of the ASE calling dataset and SNP annotation2 -the output of
getIntersection()}
}
\value{
a dataframe object with SNPs common to df_aseSnpAnn1 and
df_aseSnpAnn2
}
\description{
The function takes as input 2 data frames, one with the
intersection of the ASE calling dataset and the SNP annotation1
and the second the intersection of the ASE calling and SNP
annotation2 dataset. These can be obtained by calling the
getIntersection(). Each of the data frame objects must have
a column named 'cmp.col', e.g. containing
the rsid, chr:pos values, that is used to merge the data.
}
