% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integration}
\alias{integration}
\title{Runs Integration step and calculates the p-value}
\usage{
integration(rdaSnpAnn1, rdaSnpAnn2, alphaVal = 0.5,
  nIterations = 10000, outFilePrefix, seedValue = NULL)
}
\arguments{
\item{rdaSnpAnn1}{the transformed zscores rda file name
generated by the randomisation step-MAE for the
SNPannotation1
- *obsRandomValsMean_zscoreNmlDistr.rda}

\item{rdaSnpAnn2}{the transformed zscores rda file name
generated by the randomisation step-MAE for the
SNPannotation2
- *obsRandomValsMean_zscoreNmlDistr.rda}

\item{alphaVal}{numeric data type with values
from 0-1. default = 0.5}

\item{nIterations}{number of random selction iterations
default =10000}

\item{outFilePrefix}{The names of all the output files
generated will be assigned this prefix}

\item{seedValue}{value of the seed to be set.
default = NULL = No seed set}
}
\value{
the pnorm p-value for the alpha in parameter
alphaVal
}
\description{
Runs Integration step and calculates the p-value
}
